Reworked installcert
