+++ kernelerror/tools/misc/installcert.py
@@ -1,0 +1,36 @@
+import os
+from drozer.modules import common, Module
+from drozer import android
+import argparse
+
+class InstallCert(Module):
+
+    name = "Install a CA cert"
+    description = ""
+    author = "Bernard Wagner"
+    date = "2014-09-22"
+    license = "BSD (3-clause)"
+    path = ["tools", "misc"]
+
+    def add_arguments(self,parser):
+        parser.add_argument("source",help="specify the path of the certificate to install")
+        parser.add_argument("-n","--name",default=None,help="specify the name of the certificate")
+
+    def execute(self, arguments):
+        try:
+            print "Uploading"
+            myfile = open(arguments.source, 'rb')
+            with myfile:
+                cert = myfile.read()
+                intent = android.Intent(action="android.credentials.INSTALL", extras = [], flags = [])
+                name = arguments.name 
+                if arguments.name == None:
+                    name = os.path.basename(arguments.source).split('.')[0]
+                intent.extras.append(("string","name",name))
+                intent.extras.append(("bytearray","CERT",cert))
+                intent.flags.append(("ACTIVITY_NEW_TASK"))
+                self.getContext().startActivity(intent.buildIn(self))
+                print "Done"
+        except Exception as error:
+            print error
+
